--- src/Aydsko.iRacingData.UnitTests/Responses/ValidCredentialsIsSuccessfulAsync/0.json
@@ -1,14 +1,0 @@
-﻿{
-  "headers": {},
-  "content": {
-    "authcode": "51687dfd-1d8e-4faa-bf72-52cb36e4f762",
-    "autoLoginSeries": null,
-    "autoLoginToken": null,
-    "custId": 123456,
-    "email": "test.user@example.com",
-    "ssoCookieDomain": ".iracing.com",
-    "ssoCookieName": "irsso_membersv2",
-    "ssoCookiePath": "/",
-    "ssoCookieValue": "3883fc6a-890c-4c75-981c-84f2f9ebfb41"
-  }
-}
