+++ src/Aydsko.iRacingData/Aydsko.iRacingData.csproj
@@ -65,6 +65,16 @@
       <_Parameter3_IsLiteral>true</_Parameter3_IsLiteral>
       <_Parameter4>Scope = "type"</_Parameter4>
       <_Parameter4_IsLiteral>true</_Parameter4_IsLiteral>
+      <_Parameter5>Target = "~T:Aydsko.iRacingData.iRacingClientOptionsValueMissingException"</_Parameter5>
+      <_Parameter5_IsLiteral>true</_Parameter5_IsLiteral>
+    </AssemblyAttribute>
+    <AssemblyAttribute Include="System.Diagnostics.CodeAnalysis.SuppressMessageAttribute">
+      <_Parameter1>Style</_Parameter1>
+      <_Parameter2>IDE1006:Naming Styles</_Parameter2>
+      <_Parameter3>Justification = "&lt;The company name which supplies this data API uses the lowercase i prefix so this is a proper name.&gt;"</_Parameter3>
+      <_Parameter3_IsLiteral>true</_Parameter3_IsLiteral>
+      <_Parameter4>Scope = "type"</_Parameter4>
+      <_Parameter4_IsLiteral>true</_Parameter4_IsLiteral>
       <_Parameter5>Target = "~T:Aydsko.iRacingData.iRacingDataClientOptions"</_Parameter5>
       <_Parameter5_IsLiteral>true</_Parameter5_IsLiteral>
     </AssemblyAttribute>
