+++ src/Aydsko.iRacingData/ServicesExtensions.cs
@@ -9,26 +9,32 @@ namespace Aydsko.iRacingData;
 
 public static class ServicesExtensions
 {
-    public static IServiceCollection AddIRacingDataApi(this IServiceCollection services)
-    {
-        return AddIRacingDataApi(services, (options) => { });
-    }
-
-    public static IServiceCollection AddIRacingDataApi(this IServiceCollection services, Action<iRacingDataClientOptions> configureOptions)
+    public static IServiceCollection AddIRacingDataApi(this IServiceCollection services!!, Action<iRacingDataClientOptions> configureOptions!!)
     {
         services.TryAddSingleton(new CookieContainer());
 
         var options = new iRacingDataClientOptions();
 
-        if (configureOptions is not null)
+#pragma warning disable CA1062 // Validate arguments of public methods
+        configureOptions(options);
+#pragma warning restore CA1062 // Validate arguments of public methods
+
+        if (string.IsNullOrWhiteSpace(options.Username))
         {
-            configureOptions(options);
+            throw iRacingClientOptionsValueMissingException.Create(nameof(options.Username));
         }
 
-        services.AddSingleton(options);
+        if (string.IsNullOrWhiteSpace(options.Password))
+        {
+            throw iRacingClientOptionsValueMissingException.Create(nameof(options.Password));
+        }
 
-        services.AddHttpClient<iRacingDataClient>()
-                .ConfigurePrimaryHttpMessageHandler(services => new HttpClientHandler {
+        services.AddHttpClient<IDataClient, DataClient>((HttpClient httpClient, IServiceProvider provider) => new DataClient(httpClient,
+                                                                                                                                                  provider.GetRequiredService<ILogger<DataClient>>(),
+                                                                                                                                                  options,
+                                                                                                                                                  provider.GetRequiredService<CookieContainer>()))
+                .ConfigurePrimaryHttpMessageHandler(services => new HttpClientHandler
+                {
                     UseCookies = true,
                     CookieContainer = services.GetRequiredService<CookieContainer>()
                 });
@@ -36,4 +42,3 @@ public static class ServicesExtensions
         return services;
     }
 }
-
